+++ pip_audit/_dependency_source/pyproject.py
@@ -76,7 +76,13 @@ class PyProjectSource(DependencySource):
             # dependency resolution now, we can think about doing `pip install <local-project-dir>`
             # regardless of whether the project has a `pyproject.toml` or not. And if it doesn't
             # have a `pyproject.toml`, we can raise an error if the user provides `--fix`.
-            with NamedTemporaryFile() as req_file, TemporaryDirectory() as ve_dir:
+            with TemporaryDirectory() as ve_dir, NamedTemporaryFile(
+                dir=ve_dir, delete=False
+            ) as req_file:
+                # We use delete=False in creating the tempfile to allow it to be
+                # closed and opened multiple times within the context scope on
+                # windows, see GitHub issue #646.
+
                 # Write the dependencies to a temporary requirements file.
                 req_file.write(os.linesep.join(deps).encode())
                 req_file.flush()